Enforce grading core isolation
