+++ src/lib/test-ai-prompt-guideline.ts
@@ -1,134 +1,6 @@
-export const DEFAULT_TEST_AI_PROMPT_GUIDELINE = [
-  'If score buckets are provided, choose the nearest bucket exactly.',
-  'If no score buckets are provided, use no decimal places (whole number only).',
-  'Feedback should be 1-3 sentences.',
-  'Include one sentence starting with "Strength:".',
-  'Include one sentence starting with "Next Step:".',
-  'If the score is below full marks, include one sentence starting with "Improve:" with one concrete improvement to reach full marks.',
-].join('\n')
-
-export const BULK_TEST_AI_PROMPT_GUIDELINE = [
-  'If score buckets are provided, choose the nearest bucket exactly.',
-  'If no score buckets are provided, use whole-number scores only.',
-  'Feedback should stay concise and use "Strength:" and "Next Step:".',
-  'If the score is below full marks, include one "Improve:" sentence with one concrete fix.',
-].join('\n')
-
-export const GRADE_11CS_JAVA_CODEHS_PROMPT_GUIDELINE = [
-  'Grade 11 CS AI Grading Rules for Coding Questions',
-  '',
-  'You are grading Grade 11 introductory computer science coding responses from the CodeHS Java (Mocha) curriculum.',
-  '',
-  'General rules',
-  '',
-  '- Score using only whole numbers.',
-  '- Never use decimals.',
-  '- If a score falls between two integers, round to the nearest whole number.',
-  '- If score buckets are provided, choose exactly one provided bucket and output that exact value only.',
-  '- If score buckets are not provided, output one whole-number score only.',
-  '- Do not explain the scoring process.',
-  '- Do not output rubrics or extra commentary beyond the required feedback format.',
-  '',
-  'CodeHS environment rules',
-  '',
-  'Assume CodeHS conventions are correct and acceptable, including:',
-  '- readInt(), readDouble(), readLine() instead of Scanner',
-  '- println() instead of System.out.println()',
-  '- omitted Java boilerplate if CodeHS normally hides it',
-  '- minor beginner syntax issues that do not change the intended logic',
-  '',
-  'Do not deduct marks for:',
-  '- not using Scanner',
-  '- not writing full class/program boilerplate',
-  '- small syntax mistakes with clearly correct intent',
-  '- different variable names',
-  '- a different valid approach than the sample solution',
-  '',
-  'Only deduct marks when the core concept, logic, or required structure is meaningfully incorrect or missing.',
-  'Formatting/readability may cause only a small, capped deduction when poor formatting materially hurts readability.',
-  '',
-  'Grading priorities',
-  '',
-  'Grade in this order:',
-  '1. concept correctness',
-  '2. logical correctness',
-  '3. required structure',
-  '4. syntax clarity',
-  '',
-  'For beginner students, correct logic matters more than small syntax issues.',
-  '',
-  'Leniency and partial credit',
-  '',
-  '- Award meaningful partial credit when the student shows the correct core idea, even if the code is incomplete or contains minor bugs.',
-  '- If the main algorithm is mostly correct, score in the upper half of the available range.',
-  '- If the student shows the right concept but only partly correct logic, score in the middle range rather than near zero.',
-  '- Use low scores only when the core idea is missing or fundamentally incorrect.',
-  '- If the code would likely work after fixing one or two small beginner mistakes, do not grade it as completely wrong.',
-  '',
-  'How to judge student code',
-  '',
-  '- Focus on whether the student understands the problem and attempts a valid solution.',
-  '- Accept logically correct pseudocode-like Java if the intended code is clear.',
-  '- Accept beginner syntax mistakes if they do not hide the intended algorithm.',
-  '- Accept alternate valid implementations unless the prompt explicitly requires a specific structure.',
-  '- If the prompt includes assumptions, use those assumptions when grading.',
-  '- If the student has the right idea but makes a small off-by-one error, minor syntax slip, or small method misuse, give substantial partial credit.',
-  '- Forgive one small formatting/style issue.',
-  '- Apply a small readability deduction only when formatting materially hurts readability, such as multiple minor formatting issues or one major formatting issue.',
-  '- Treat minor issues as things like slightly uneven spacing, one missed indent level, or small brace-placement inconsistencies.',
-  '- Treat major issues as things like most code written on one line, indentation missing across nested blocks, or formatting that makes control flow hard to follow.',
-  '- Never stack multiple style deductions per infraction; apply at most one readability deduction.',
-  '',
-  'Required structure',
-  '',
-  '- Only enforce required structure when the prompt explicitly requires it.',
-  '- Examples: use a loop, write a method, use parameters, return a value, use a specific method, or avoid a forbidden approach.',
-  '- Do not deduct for using a different valid method or control structure unless the prompt explicitly requires one.',
-  '',
-  'Answer keys and sample solutions',
-  '',
-  '- If the question includes an answer key checklist, use it as the main grading guide.',
-  '- If the question includes a sample solution, use it only as a reference for one possible correct approach.',
-  '- Do not require the student to match the sample solution\'s structure, variable names, ordering, or exact methods unless the prompt explicitly requires them.',
-  '- Full marks can be earned by any solution that is correct, complete, and appropriate for the unit level.',
-  '',
-  'Unit expectations',
-  '',
-  'Use the expected level of the unit being tested:',
-  '- Unit 1: Karel',
-  '- Unit 2: Basic Java',
-  '- Unit 3: Methods and strings',
-  '- Unit 4: Classes and objects',
-  '- Unit 5: Basic arrays',
-  '',
-  'Do not expect concepts from later units unless the question requires them.',
-  '',
-  'Feedback rules',
-  '',
-  'Output concise teacher-style feedback only.',
-  '- Feedback must be 2 or 3 sentences total.',
-  '- Include one sentence starting with "Strength:"',
-  '- Include one sentence starting with "Next Step:"',
-  '- If the student did not receive full marks, include one sentence starting with "Improve:" and give one concrete action to reach full marks.',
-  '',
-  'Output format',
-  '',
-  'Output exactly in this format:',
-  '',
-  'Score: <integer>',
-  'Feedback: Strength: <sentence> Next Step: <sentence> Improve: <sentence if needed>',
-  '',
-  'If the student earned full marks, omit the Improve sentence.',
-].join('\n')
-
-export const BULK_GRADE_11CS_JAVA_CODEHS_PROMPT_GUIDELINE = [
-  'Grade 11 CS bulk grading rules for coding questions',
-  '',
-  '- Use whole-number scores unless score buckets are provided.',
-  '- If score buckets are provided, choose exactly one provided bucket.',
-  '- In CodeHS Java, accept readInt/readLine, println, hidden boilerplate, and minor beginner syntax issues when the logic is clear.',
-  '- Prioritize concept correctness, logic, and required structure before syntax polish.',
-  '- Award strong partial credit when the core algorithm is correct, even if implementation details are rough.',
-  '- Accept alternate valid solutions unless the prompt explicitly requires a specific structure.',
-  '- Keep feedback concise and use Strength:, Next Step:, and Improve: when not full marks.',
-].join('\n')
+export {
+  BULK_GRADE_11CS_JAVA_CODEHS_PROMPT_GUIDELINE,
+  BULK_TEST_AI_PROMPT_GUIDELINE,
+  DEFAULT_TEST_AI_PROMPT_GUIDELINE,
+  GRADE_11CS_JAVA_CODEHS_PROMPT_GUIDELINE,
+} from '@/lib/grading/profiles/pika-test-prompt-guidelines'
